Complete rewrite of letter rendering and pdf generation using queue and threading.
This was really a lot of work, but I think the code is *much* cleaner now.

The LetterExportAssistant now only queues the contracts and updates the GUI. Composing and Prerendering is done in core.letterrenderer (using threading).

Furthermore, Invoices and Notes have been added to the database as LetterParts (many-to-many relationship), so we can later use them to store correspondence. This breaks compatability with old DB revisions.

LetterComposition is now a separate class in core.lettercomposition.